+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Now accept non-positive integers.
   - Changed default values to `0` and `-2` respectively.
 - Swapped `N` for `A` in the *method* field of the `iterm2` style-speific render format specification ([#92]).
+- `StyleError` is now raised instead of style-specific exceptions ([#93]).
 
 ### Removed
 - Image scaling ([#88]).
@@ -52,12 +53,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Replaced by terminal size-relative frame sizes ([#89]) and padding ([#91]).
 - *native* and *stall_native* `iterm2` style-specific render parameters ([#92]).
   - Replaced by the `ANIM` render method.
+- Style-specific exceptions ([#93]).
+  - `GraphicsImageError`
+  - `TextImageError`
+  - `BlockImageError`
+  - `ITerm2ImageError`
+  - `KittyImageError`
 
 [#87]: https://github.com/AnonymouX47/term-image/pull/87
 [#88]: https://github.com/AnonymouX47/term-image/pull/88
 [#89]: https://github.com/AnonymouX47/term-image/pull/89
 [#91]: https://github.com/AnonymouX47/term-image/pull/91
 [#92]: https://github.com/AnonymouX47/term-image/pull/92
+[#93]: https://github.com/AnonymouX47/term-image/pull/93
 [08f4e4d]: https://github.com/AnonymouX47/term-image/commit/08f4e4d1514313bbd4278dadde46d21d0b11ed1f
 [fa47742]: https://github.com/AnonymouX47/term-image/commit/fa477424c83474256d4972c4b2cdd4a765bc1cda
 [ed3baa3]: https://github.com/AnonymouX47/term-image/commit/ed3baa38d7621720c007f4662f89d7abadd76ec3